requirements.txt
pytest==8.3.5
allure-pytest==2.15.0
selenium==4.20.0
webdriver-manager==4.0.1
